+++ aws_lambda_powertools/event_handler/openapi/types.py
@@ -1,0 +1,3 @@
+from typing import Any, Callable, Optional
+
+CacheKey = Optional[Callable[..., Any]]
